requirements.txt
contentful==1.13.1
python-dotenv==1.0.0